Directory line rendering
