+++ npm/design-system/src/components/fileTree/FileTree.tsx
@@ -5,13 +5,9 @@ import { CollapsibleGroupHeader } from 'components/collapsibleGroup/CollapsibleG
 import { LeafProps, OnNodePress, ParentProps } from 'components/virtualizedTree/types'
 import { buildTree } from './buildTree'
 import { FileBase, FilePressEvent, FileTreeProps, TreeFile, TreeFolder } from './types'
+import { FileTreeFile } from './FileTreeFile'
 
-import type { IconifyIcon } from '@iconify/types'
-import { InlineIcon } from '@iconify/react'
-import javascriptIcon from '@iconify/icons-vscode-icons/file-type-js-official'
-import typescriptIcon from '@iconify/icons-vscode-icons/file-type-typescript-official'
-import reactJs from '@iconify/icons-vscode-icons/file-type-reactjs'
-import reactTs from '@iconify/icons-vscode-icons/file-type-reactts'
+import styles from './FileTree.module.scss'
 
 interface MutableFilePressEvent extends Omit<FilePressEvent, 'defaultPrevented'> {
   defaultPrevented: boolean
@@ -64,8 +60,6 @@ export const FileTree = <T extends FileBase>({
     <VirtualizedTree<TreeFile<T>, TreeFolder<T>>
       tree={tree}
       defaultItemSize={20}
-      // TODO: Set off of design-system spacing
-      indentSize={1}
       showRoot={true}
       onNodePress={onNodePress}
       onRenderParent={ParentComponent}
@@ -74,63 +68,27 @@ export const FileTree = <T extends FileBase>({
   )
 }
 
-const DefaultFolder = <T extends FileBase>({ parent, isOpen }: ParentProps<TreeFolder<T>>) => (
+const DefaultFolder = <T extends FileBase>({ parent: { name }, depth, isOpen }: ParentProps<TreeFolder<T>>) => (
   <CollapsibleGroupHeader
-    title={parent.name}
+    className={styles.node}
+    style={depth > 0 ? {
+      paddingLeft: `${depth}rem`,
+      backgroundSize: `${depth}rem 100%`,
+    } : undefined}
+    title={name}
     expanded={isOpen}
     icons={{ expanded: 'chevron-down', collapsed: 'chevron-right' }}
   />
 )
 
 const DefaultFile = <T extends FileBase>(props: LeafProps<TreeFile<T>>) => (
-  <FileLeaf
+  <FileTreeFile
     {...props}
+    style={props.depth > 0 ? {
+      paddingLeft: `${props.depth}rem`,
+      backgroundSize: `${props.depth}rem 100%`,
+    } : undefined}
     item={props.leaf}
     indexes={[]}
   />
 )
-
-export interface NodeComponentProps<T> {
-  item: T
-  depth: number
-  indexes: number[]
-}
-
-export interface FileComponentProps<T extends FileBase> extends NodeComponentProps<TreeFile<T>> {
-  depth: number
-  remeasure: () => void
-}
-
-export const icons: Record<string, { icon: IconifyIcon }> = {
-  js: { icon: javascriptIcon },
-  ts: { icon: typescriptIcon },
-  tsx: { icon: reactTs },
-  jsx: { icon: reactJs },
-} as const
-
-const FileLeaf = <T extends FileBase>({ item, indexes }: FileComponentProps<T>) => {
-  const ext = getExt(item.name)
-  const inlineIconProps = ext ? icons[ext] : {
-    // If we don't have an icon for the extension, don't render an icon
-    icon: '',
-  }
-
-  return (
-    <div>
-      <InlineIcon {...inlineIconProps} />
-      {/* <NameWithHighlighting
-        item={item}
-        indexes={indexes}
-      /> */}
-      {item.name}
-    </div>
-  )
-}
-
-const extensionRegex = /(?:\.([^.]+))?$/
-
-export const getExt = (path: string) => {
-  const extensionMatches = path.match(extensionRegex)
-
-  return extensionMatches ? extensionMatches[1] : ''
-}
