+++ npm/design-system/src/css/semanticColors.scss
@@ -19,6 +19,8 @@
 $black-rgb-color: var(--black-rgb-color);
 
 $control-border-color: $metal-30;
+// TODO: Find a better name
+$control-slight-color: $metal-10;
 $control-focus-color: $chill-40;
 
 $button-blue-color: $chill-50;
